Put two console test files back together
